requirements.txt
pyfakefs
pytest
hypothesis
click
python-crontab
